utils/requirements.txt
annoy==1.9.3
Pillow==4.1.1
nltk==3.2.3
numpy==1.12.1
scikit-learn==0.18.1
scipy==0.19.0
